perf: finalize async I/O in storage and fix Rust 2024 compliance
- Replaced synchronous file I/O with tokio::fs across all initialization paths.
- Added parent directory creation logic for robust startup.
- Fixed Rust 2024 compliance by wrapping std::env calls in unsafe blocks.
- Added #[allow(unsafe_code)] to test modules using environment manipulation.
- Integrated serial_test to prevent race conditions in environment-dependent tests.
- Refactored tests to return Result and avoid unwrap/expect as per project lints.
- Fixed target dependency syntax in Cargo.toml.
