--- memory-storage-turso/src/compression/payload_tests.rs
@@ -1,133 +1,0 @@
-use super::*;
-use crate::compression::CompressionAlgorithm;
-
-#[test]
-fn test_compress_threshold_boundary() {
-    let data = b"a".repeat(100);
-
-    // Below threshold: should be None
-    let compressed = CompressedPayload::compress(&data, 101).unwrap();
-    assert_eq!(compressed.algorithm, CompressionAlgorithm::None);
-    assert_eq!(compressed.original_size, 100);
-    assert_eq!(compressed.compressed_size, 100);
-    assert_eq!(compressed.compression_ratio, 1.0);
-
-    // At threshold: should compress (100 < 100 is false)
-    let compressed = CompressedPayload::compress(&data, 100).unwrap();
-    assert_ne!(compressed.algorithm, CompressionAlgorithm::None);
-
-    // Above threshold: should compress
-    let compressed = CompressedPayload::compress(&data, 50).unwrap();
-    assert_ne!(compressed.algorithm, CompressionAlgorithm::None);
-}
-
-#[test]
-fn test_bandwidth_savings_percent() {
-    let payload = CompressedPayload {
-        original_size: 100,
-        compressed_size: 60,
-        compression_ratio: 0.6,
-        data: vec![],
-        algorithm: CompressionAlgorithm::Zstd,
-    };
-    assert_eq!(payload.bandwidth_savings_percent(), 40.0);
-
-    let payload_none = CompressedPayload {
-        original_size: 100,
-        compressed_size: 100,
-        compression_ratio: 1.0,
-        data: vec![],
-        algorithm: CompressionAlgorithm::None,
-    };
-    assert_eq!(payload_none.bandwidth_savings_percent(), 0.0);
-}
-
-#[test]
-fn test_roundtrip_none() {
-    let data = b"small data";
-    let compressed = CompressedPayload::compress(data, 1024).unwrap();
-    assert_eq!(compressed.algorithm, CompressionAlgorithm::None);
-
-    let decompressed = compressed.decompress().unwrap();
-    assert_eq!(data.to_vec(), decompressed);
-}
-
-#[test]
-#[cfg(feature = "compression-lz4")]
-fn test_lz4_explicit_roundtrip() {
-    let data = b"repetitive data".repeat(50);
-    let compressed = CompressedPayload::compress_lz4(&data).unwrap();
-    assert_eq!(compressed.algorithm, CompressionAlgorithm::Lz4);
-    assert!(compressed.compression_ratio < 1.0);
-
-    let decompressed = compressed.decompress().unwrap();
-    assert_eq!(data, decompressed);
-}
-
-#[test]
-#[cfg(feature = "compression-zstd")]
-fn test_zstd_explicit_roundtrip() {
-    let data = b"repetitive data".repeat(50);
-    let compressed = CompressedPayload::compress_zstd(&data).unwrap();
-    assert_eq!(compressed.algorithm, CompressionAlgorithm::Zstd);
-    assert!(compressed.compression_ratio < 1.0);
-
-    let decompressed = compressed.decompress().unwrap();
-    assert_eq!(data, decompressed);
-}
-
-#[test]
-#[cfg(feature = "compression-gzip")]
-fn test_gzip_explicit_roundtrip() {
-    let data = b"repetitive data".repeat(50);
-    let compressed = CompressedPayload::compress_gzip(&data).unwrap();
-    assert_eq!(compressed.algorithm, CompressionAlgorithm::Gzip);
-    assert!(compressed.compression_ratio < 1.0);
-
-    let decompressed = compressed.decompress().unwrap();
-    assert_eq!(data, decompressed);
-}
-
-#[test]
-#[cfg(all(feature = "compression-zstd", feature = "compression-lz4"))]
-fn test_compression_selection_logic() {
-    // Case 1: Highly compressible data
-    // Zstd should get ratio < 0.5.
-    // According to the current (possibly buggy) implementation:
-    // if ratio >= 0.5 { return Zstd } else { fallback to LZ4 }
-    let data = b"a".repeat(1000);
-    let compressed = CompressedPayload::compress(&data, 0).unwrap();
-
-    // Let's see what Zstd actually gives for 1000 'a's
-    let zstd = CompressedPayload::compress_zstd(&data).unwrap();
-    if zstd.compression_ratio < 0.5 {
-        // It should have fallen back to LZ4
-        assert_eq!(compressed.algorithm, CompressionAlgorithm::Lz4);
-    } else {
-        // It stayed with Zstd
-        assert_eq!(compressed.algorithm, CompressionAlgorithm::Zstd);
-    }
-
-    // Case 2: Poorly compressible data (but larger than threshold)
-    // Ratio should be >= 0.5
-    let data = b"not so repetitive data".repeat(2);
-    let compressed = CompressedPayload::compress(&data, 0).unwrap();
-    let zstd = CompressedPayload::compress_zstd(&data).unwrap();
-    if zstd.compression_ratio >= 0.5 {
-        assert_eq!(compressed.algorithm, CompressionAlgorithm::Zstd);
-    }
-}
-
-#[test]
-#[cfg(feature = "compression-zstd")]
-fn test_decompress_error_handling() {
-    // Test decompression with invalid data (if possible)
-    let payload = CompressedPayload {
-        original_size: 100,
-        compressed_size: 10,
-        compression_ratio: 0.1,
-        data: vec![0, 1, 2, 3, 4, 5], // Invalid zstd data
-        algorithm: CompressionAlgorithm::Zstd,
-    };
-    assert!(payload.decompress().is_err());
-}
